Extend transactions page

+++ PersonalFinancer.Web/Controllers/TransactionsController.cs
@@ -1,18 +1,19 @@
 ﻿namespace PersonalFinancer.Web.Controllers
 {
-    using AutoMapper;
-    using Microsoft.AspNetCore.Authorization;
-    using Microsoft.AspNetCore.Mvc;
-    using PersonalFinancer.Services.Accounts;
-    using PersonalFinancer.Services.Accounts.Models;
-    using PersonalFinancer.Services.User;
-    using PersonalFinancer.Services.User.Models;
-    using PersonalFinancer.Web.Extensions;
-    using PersonalFinancer.Web.Models.Shared;
-    using PersonalFinancer.Web.Models.Transaction;
-    using static PersonalFinancer.Data.Constants.RoleConstants;
+	using AutoMapper;
+	using Microsoft.AspNetCore.Authorization;
+	using Microsoft.AspNetCore.Mvc;
+	using PersonalFinancer.Services.Accounts;
+	using PersonalFinancer.Services.Accounts.Models;
+	using PersonalFinancer.Services.User;
+	using PersonalFinancer.Services.User.Models;
+	using PersonalFinancer.Web.Extensions;
+	using PersonalFinancer.Web.Models.Account;
+	using PersonalFinancer.Web.Models.Transaction;
+	using System.ComponentModel.DataAnnotations;
+	using static PersonalFinancer.Data.Constants.RoleConstants;
 
-    [Authorize]
+	[Authorize]
 	public class TransactionsController : Controller
 	{
 		protected readonly IAccountsUpdateService accountsUpdateService;
@@ -35,24 +36,39 @@
 		[Authorize(Roles = UserRoleName)]
 		public async Task<IActionResult> All()
 		{
-			DateTime startDate = DateTime.Now.AddMonths(-1);
-			DateTime endDate = DateTime.Now;
+			var filterDTO = new TransactionsFilterDTO
+			{
+				UserId = this.User.IdToGuid(),
+				StartDate = DateTime.Now.AddMonths(-1),
+				EndDate = DateTime.Now
+			};
 
-			UserTransactionsViewModel viewModel =
-				await this.PrepareUserTransactionsViewModel(this.User.IdToGuid(), startDate, endDate);
+			UserTransactionsViewModel viewModel = 
+				await this.PrepareUserTransactionsViewModelAsync(filterDTO);
 
 			return this.View(viewModel);
 		}
 
 		[Authorize(Roles = UserRoleName)]
 		[HttpPost]
-		public async Task<IActionResult> All(DateFilterModel inputModel)
+		public async Task<IActionResult> All(UserTransactionsInputModel inputModel)
 		{
-			if (!this.ModelState.IsValid)
-				return this.View(this.mapper.Map<UserTransactionsViewModel>(inputModel));
+			Guid userId = this.User.IdToGuid();
+			UserTransactionsViewModel viewModel;
 
-			UserTransactionsViewModel viewModel = await this.PrepareUserTransactionsViewModel(
-				this.User.IdToGuid(), inputModel.StartDate, inputModel.EndDate);
+			if (!this.ModelState.IsValid)
+			{
+				viewModel = this.mapper.Map<UserTransactionsViewModel>(inputModel);
+				viewModel.UserId = userId;
+				var dropdownDTO = await this.usersService.GetUserDropdownDataAsync(userId);
+				this.mapper.Map(dropdownDTO, viewModel);
+
+				return this.View(viewModel);
+			}
+
+			var filterDTO = this.mapper.Map<TransactionsFilterDTO>(inputModel);
+			filterDTO.UserId = userId;
+			viewModel = await this.PrepareUserTransactionsViewModelAsync(filterDTO);
 
 			return this.View(viewModel);
 		}
@@ -60,10 +76,8 @@
 		[Authorize(Roles = UserRoleName)]
 		public async Task<IActionResult> Create()
 		{
-			UserAccountsAndCategoriesServiceModel userData =
-				await this.usersService.GetUserAccountsAndCategoriesAsync(this.User.IdToGuid());
-
-			TransactionFormModel viewModel = this.mapper.Map<TransactionFormModel>(userData);
+			var viewModel = new TransactionFormViewModel() { OwnerId = this.User.IdToGuid() };
+			await this.PrepareAccountsAndCategoriesAsync(viewModel);
 			viewModel.CreatedOn = DateTime.Now;
 
 			return this.View(viewModel);
@@ -71,24 +85,22 @@
 
 		[Authorize(Roles = UserRoleName)]
 		[HttpPost]
-		public async Task<IActionResult> Create(TransactionFormModel inputModel)
+		public async Task<IActionResult> Create(TransactionFormViewModel inputModel)
 		{
+			if (inputModel.OwnerId != this.User.IdToGuid())
+				return this.BadRequest();
+
 			if (!this.ModelState.IsValid)
 			{
-				await this.PrepareTransactionFormModelForReturn(inputModel);
+				await this.PrepareAccountsAndCategoriesAsync(inputModel);
 
 				return this.View(inputModel);
 			}
 
-			if (inputModel.OwnerId != this.User.IdToGuid())
-				return this.BadRequest();
-
 			try
 			{
-				TransactionFormShortServiceModel serviceModel =
-					this.mapper.Map<TransactionFormShortServiceModel>(inputModel);
-
-				Guid newTransactionId = await this.accountsUpdateService.CreateTransactionAsync(serviceModel);
+				var dto = this.mapper.Map<CreateEditTransactionDTO>(inputModel);
+				Guid newTransactionId = await this.accountsUpdateService.CreateTransactionAsync(dto);
 				this.TempData["successMsg"] = "You create a new transaction successfully!";
 
 				return this.RedirectToAction(nameof(TransactionDetails), new { id = newTransactionId });
@@ -100,14 +112,17 @@
 		}
 
 		[HttpPost]
-		public async Task<IActionResult> Delete(Guid id, string? returnUrl = null)
+		public async Task<IActionResult> Delete([Required] Guid id, string? returnUrl = null)
 		{
+			if (!this.ModelState.IsValid)
+				return this.BadRequest();
+
 			try
 			{
 				await this.accountsUpdateService.DeleteTransactionAsync(id, this.User.IdToGuid(), this.User.IsAdmin());
 
-				this.TempData["successMsg"] = this.User.IsAdmin() ?
-					"You successfully delete a user's transaction!"
+				this.TempData["successMsg"] = this.User.IsAdmin() 
+					? "You successfully delete a user's transaction!"
 					: "Your transaction was successfully deleted!";
 			}
 			catch (ArgumentException)
@@ -124,12 +139,17 @@
 				: this.RedirectToAction("Index", "Home");
 		}
 
-		public async Task<IActionResult> TransactionDetails(Guid id)
+		public async Task<IActionResult> TransactionDetails([Required] Guid id)
 		{
+			if (!this.ModelState.IsValid)
+				return this.BadRequest();
+
+			TransactionDetailsDTO viewModel;
+
 			try
 			{
-				return this.View(await this.accountsInfoService
-					.GetTransactionDetailsAsync(id, this.User.IdToGuid(), this.User.IsAdmin()));
+				viewModel = await this.accountsInfoService
+					.GetTransactionDetailsAsync(id, this.User.IdToGuid(), this.User.IsAdmin());
 			}
 			catch (ArgumentException)
 			{
@@ -139,51 +159,48 @@
 			{
 				return this.BadRequest();
 			}
+
+			return this.View(viewModel);
 		}
 
-		public async Task<IActionResult> EditTransaction(Guid id)
+		public async Task<IActionResult> EditTransaction([Required] Guid id)
 		{
-			TransactionFormServiceModel transactionData;
+			if (!this.ModelState.IsValid)
+				return this.BadRequest();
 
 			try
 			{
-				transactionData = await this.accountsInfoService.GetTransactionFormDataAsync(id);
+				var transactionDTO = await this.accountsInfoService
+					.GetTransactionFormDataAsync(id, this.User.IdToGuid(), this.User.IsAdmin());
+
+				var viewModel = this.mapper.Map<TransactionFormViewModel>(transactionDTO);
+
+				return this.View(viewModel);
 			}
 			catch (InvalidOperationException)
 			{
 				return this.BadRequest();
 			}
-
-			if (!this.User.IsAdmin() && this.User.IdToGuid() != transactionData.OwnerId)
-				return this.Unauthorized();
-
-			TransactionFormModel viewModel = this.mapper.Map<TransactionFormModel>(transactionData);
-
-			return this.View(viewModel);
 		}
 
 		[HttpPost]
-		public async Task<IActionResult> EditTransaction(Guid id, TransactionFormModel inputModel)
+		public async Task<IActionResult> EditTransaction([Required] Guid id, TransactionFormViewModel inputModel)
 		{
-			if (!this.ModelState.IsValid)
-			{
-				await this.PrepareTransactionFormModelForReturn(inputModel);
-
-				return this.View(inputModel);
-			}
-
 			try
 			{
-				Guid ownerId = this.User.IsAdmin()
-					? await this.accountsInfoService.GetOwnerIdAsync(inputModel.AccountId ?? throw new InvalidOperationException())
-					: this.User.IdToGuid();
+				if (!this.User.IsAdmin() && this.User.IdToGuid() != inputModel.OwnerId)
+					return this.BadRequest();
 
-				if (inputModel.OwnerId != ownerId)
-					return this.Unauthorized();
+				if (!this.ModelState.IsValid)
+				{
+					await this.PrepareAccountsAndCategoriesAsync(inputModel);
 
-				TransactionFormShortServiceModel serviceModel = this.mapper.Map<TransactionFormShortServiceModel>(inputModel);
+					return this.View(inputModel);
+				}
 
-				await this.accountsUpdateService.EditTransactionAsync(id, serviceModel);
+				var dto = this.mapper.Map<CreateEditTransactionDTO>(inputModel);
+
+				await this.accountsUpdateService.EditTransactionAsync(id, dto);
 			}
 			catch (InvalidOperationException)
 			{
@@ -197,28 +214,26 @@
 			return this.RedirectToAction(nameof(TransactionDetails), new { id });
 		}
 
-		private async Task PrepareTransactionFormModelForReturn(TransactionFormModel formModel)
+		/// <summary>
+		/// Throws InvalidOperationException when Owner ID is invalid.
+		/// </summary>
+		/// <exception cref="InvalidOperationException"></exception>
+		private async Task PrepareAccountsAndCategoriesAsync(TransactionFormViewModel formModel)
 		{
-			if (formModel.OwnerId != null)
-			{
-				UserAccountsAndCategoriesServiceModel userData =
-					await this.usersService.GetUserAccountsAndCategoriesAsync(
-						formModel.OwnerId ?? throw new InvalidOperationException());
-
-				formModel.UserCategories = userData.UserCategories;
-				formModel.UserAccounts = userData.UserAccounts;
-			}
+			Guid ownerId = formModel.OwnerId ?? throw new InvalidOperationException("Owner ID cannot be null.");
+			var accountsAndCategoriesDTO = await this.usersService.GetUserAccountsAndCategoriesDropdownDataAsync(ownerId);
+			this.mapper.Map(accountsAndCategoriesDTO, formModel);
 		}
 
-		private async Task<UserTransactionsViewModel> PrepareUserTransactionsViewModel(
-			Guid userId, DateTime startDate, DateTime endDate)
+		private async Task<UserTransactionsViewModel> PrepareUserTransactionsViewModelAsync(
+			TransactionsFilterDTO filterDTO)
 		{
-			TransactionsServiceModel userTransactions =
-				await this.accountsInfoService.GetUserTransactionsAsync(userId, startDate, endDate);
-
-			UserTransactionsViewModel viewModel = this.mapper.Map<UserTransactionsViewModel>(userTransactions);
-			viewModel.Id = userId;
-			viewModel.Pagination.TotalElements = userTransactions.TotalTransactionsCount;
+			TransactionsPageDTO transactionsDTO =
+				await this.usersService.GetUserTransactionsPageDataAsync(filterDTO);
+			var viewModel = new UserTransactionsViewModel(
+				transactionsDTO.TotalTransactionsCount, filterDTO.Page);
+			this.mapper.Map(filterDTO, viewModel);
+			this.mapper.Map(transactionsDTO, viewModel);
 
 			return viewModel;
 		}
